backend/requirements.txt
# Core Framework
fastapi==0.122.0
uvicorn==0.38.0
starlette==0.50.0

# Database
sqlalchemy==2.0.44
alembic==1.17.2

# Pydantic
pydantic==2.12.5
pydantic-settings==2.12.0

# Authentication - استخدام python-jose فقط (إزالة التكرار)
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# API Utils
python-multipart==0.0.20
slowapi==0.1.9
python-dotenv==1.2.1

# Email
fastapi-mail==1.6.0

# PDF & QR
reportlab==4.4.6
qrcode==8.2
Pillow==12.0.0

# Testing
pytest==9.0.2
httpx==0.28.1
